Get rid of compilation warnings.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@53327 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ src/connect/ncbi_namedpipe.cpp
@@ -497,7 +497,7 @@ EIO_Status CNamedPipeHandle::Read(void* buf, size_t count, size_t* n_read,
             // We must read only "count" bytes of data regardless of the amount
             // available to read.
             if (bytes_avail > count) {
-                bytes_avail = count;
+                bytes_avail = (DWORD)count;
             }
             if ( !::ReadFile(m_Pipe, buf, bytes_avail, &bytes_avail, NULL) ) {
                 if ( !bytes_avail ) {
